Cleanup and reorg so that Value doesn't know about Syntax
