+++ src/Abstract/Interpreter.hs
@@ -1,13 +1,13 @@
 {-# LANGUAGE UndecidableInstances, AllowAmbiguousTypes, ConstraintKinds, DataKinds, FlexibleContexts, FlexibleInstances, MultiParamTypeClasses, ScopedTypeVariables, TypeApplications, TypeOperators #-}
 module Abstract.Interpreter where
 
+import Abstract.Environment
 import Abstract.Primitive
 import Abstract.Store
 import Abstract.Syntax
 import Abstract.Term
 import Abstract.Type
-import Abstract.Value
-import Abstract.Environment
+
 import Control.Effect
 import Control.Monad.Effect hiding (run)
 import Control.Monad.Effect.Fail
@@ -27,7 +27,7 @@ type EvalResult l v = Final (Interpreter l v) v
 
 
 eval' :: forall l v . (Ord v, Eval v (Eff (Interpreter l v)) Syntax Syntax, MonadAddress l (Eff (Interpreter l v)), MonadPrim v (Eff (Interpreter l v)), Semigroup (Cell l v))
-     => Term Prim Syntax
+     => Term Syntax
      -> EvalResult l v
 eval' = run @(Interpreter l v) . fix (\ ev -> eval ev . out)
 -- eval' = run @(Interpreter l v) . fix (ev @l)
