+++ src/Abstract/Term.hs
@@ -1,29 +1,29 @@
-{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE TypeOperators, MultiParamTypeClasses #-}
 
 module Abstract.Term where
 
-import Data.Union
-import Data.Functor.Classes
 import Data.Function
+import Data.Functor.Classes
+import Data.Union
 
 type Name = String
 
--- a - Prim
--- f - Syntax
-newtype Term a f = In { out :: f (Term a f)}
+newtype Term syntax = In { out :: syntax (Term syntax)}
+
+class Monad m => Eval v m g syntax where
+  eval :: (Term g -> m v) -> syntax (Term g) -> m v
 
 -- Smart constructor helper for Term
-inject :: (g :< fs) => g (Term a (Union fs)) -> Term a (Union fs)
+inject :: (g :< fs) => g (Term (Union fs)) -> Term (Union fs)
 inject = In . inj
-
 
 -- Instances
 
-instance Eq1 syntax => Eq (Term a syntax) where
+instance Eq1 syntax => Eq (Term syntax) where
   (==) = liftEq (==) `on` out
 
-instance Ord1 syntax => Ord (Term a syntax) where
+instance Ord1 syntax => Ord (Term syntax) where
   compare (In a) (In b) = liftCompare compare a b
 
-instance Show1 syntax => Show (Term a syntax) where
+instance Show1 syntax => Show (Term syntax) where
   showsPrec p = liftShowsPrec showsPrec showList p . out
